Install playwright
