+++ packages/cookie-banner/example/src/index.html
@@ -1,60 +1,72 @@
 <!doctype html>
-<html>
-  <head>
+<html lang="en">
+<head>
     <title>StormID</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <style>
         * {
-            margin:0;
-            padding:0;
-            box-sizing:border-box;
+            margin: 0;
+            padding: 0;
+            box-sizing: border-box;
             font-family: sans-serif;
         }
+
         html,
         body {
-            height:100%;
+            height: 100%;
         }
+
         hr {
-            height:1px;
-            width:100%;
-            display:block;
-            background:#ddd;
-            margin:24px 0;
-            border:0 none;
+            height: 1px;
+            width: 100%;
+            display: block;
+            background: #ddd;
+            margin: 24px 0;
+            border: 0 none;
         }
+
         h1 {
             font: 1.25rem/1.2 sans-serif;
             margin-top: 1.5rem;
         }
+
         h2 {
             font: 1.1rem/1.2 sans-serif;
-            margin-top:12px;
-            padding-top:12px;
+            margin-top: 12px;
+            padding-top: 12px;
         }
+
         h3 {
             font: 1rem/1.2 sans-serif;
-            margin-top:12px;
-            padding-top:12px;
+            margin-top: 12px;
+            padding-top: 12px;
         }
+
         p {
-            margin-top:1.5rem;
+            margin-top: 1.5rem;
         }
+
         fieldset {
             padding: 0;
             margin: 0;
             bordeR: 0 none;
         }
+
         legend {
             display: block;
         }
+
         .container {
             margin: 24px auto 48px;
             width: calc(100% - 48px);
         }
+
         .embed iframe,
         .embed__placeholder {
             width: 600px;
             aspect-ratio: 16 / 9;
         }
+
         .embed__placeholder {
             background-color: hsl(0, 0%, 90%);
             display: flex;
@@ -63,6 +75,7 @@
             align-items: center;
             gap: 12px;
         }
+
         .privacy-banner__btn-text,
         .embed__placeholder-btn,
         .privacy-banner__link {
@@ -72,65 +85,79 @@
             font-size: 1rem;
             padding: 0;
         }
+
         .privacy-banner__title {
-            font-size:1.2rem;
+            font-size: 1.2rem;
             margin-bottom: 12px;
         }
+
         .privacy-banner {
             position: fixed;
             bottom: 0;
-            left:0;
-            right:0;
-            padding:24px 25px;
+            left: 0;
+            right: 0;
+            padding: 24px 25px;
             /* border-top:1px solid #191919; */
             z-index: 1000;
-            box-shadow: 0 1px 3px 1px rgba(60,64,67,.2), 0 2px 8px 4px rgba(60,64,67,.1);
+            box-shadow: 0 1px 3px 1px rgba(60, 64, 67, .2), 0 2px 8px 4px rgba(60, 64, 67, .1);
             background-color: #fff;
         }
+
         .privacy-banner__list {
             list-style: none;
             margin-bottom: 24px;
         }
+
         button,
         .privacy-banner__btn {
-            background-color:#191919;
-            padding:8px 16px;
+            background-color: #191919;
+            padding: 8px 16px;
             color: #fff;
-            border:0 none;
+            border: 0 none;
             cursor: pointer;
             display: inline-block;
         }
-        button + button {
+
+        button+button {
             margin-left: 12px;
         }
+
         .privacy-banner__submit:hover,
         .privacy-banner__btn:hover {
             background-color: blue
         }
+
         .privacy-banner__set-accept {
             display: inline-block;
         }
+
         .privacy-banner__actions {
             padding-top: 24px;
         }
+
         .privacy-banner__row {
             margin-top: 12px;
         }
+
         .privacy-banner__legend {
             padding-top: 12px;
         }
+
         .privacy-banner__form-title {
             font-size: 1.2rem;
         }
+
         .privacy-banner__form-description {
             display: block;
             padding-top: 12px;
         }
+
         .privacy-banner__label-description {
             display: block;
             padding-left: 18px;
             font-size: .85rem;
         }
+
         .privacy-banner__form-announcement {
             position: absolute;
             opacity: 0;
@@ -142,35 +169,46 @@
             clip: rect(0 0 0 0);
             border: 0;
         }
+
         .privacy-banner__panel {
-            display:none;
+            display: none;
         }
+
         .privacy-banner__submit {
             margin-top: 12px;
         }
+
         .on--preferences .privacy-banner__panel {
-            display:block;
+            display: block;
         }
+
         .on--preferences .privacy-banner__summary,
         .on--preferences .privacy-banner__actions {
             display: none;
         }
     </style>
-  </head>
-  <body>
-        <div class="container">
-            <div class="embed">
-                <div class="embed__placeholder" data-iframe-src="https://www.youtube.com/embed/qpLKTUQev30" data-iframe-title="test video">
-                    <p>Update your cookie preferences to view this content.</p>
-                    <button type="button" class="embed__placeholder-btn js-preferences-update">Update preferences</button>
-                </div>
+</head>
+
+<body>
+    <main class="container">
+        <div class="embed">
+            <div class="embed__placeholder" data-iframe-src="https://www.youtube.com/embed/qpLKTUQev30"
+                data-iframe-title="test video">
+                <p>Update your cookie preferences to view this content.</p>
+                <button type="button" class="embed__placeholder-btn js-preferences-update">Update preferences</button>
             </div>
-            <div class="privacy-banner__update"></div>
-            <h1>Cookies</h1>
-            <p>Cookies are a common feature used on almost all websites. A cookie is a small text file that is downloaded onto a device to allow a website to recognise it and to store some information about your preferences.</p>
-            <p>We use cookies for different reasons, including for managing your shopping cart, measuring your visits on our sites, recognising and remembering your preferences and showing you personalised ads. Learn more about cookies and how we use them.</p>
-            <div class="privacy-banner__form-container "></div>
-            <div class="privacy-banner__form-announcement" style="position: absolute;width:1px;height:1px;opacity:0;" role="alert"></div>
         </div>
-  </body>
+        <div class="privacy-banner__update"></div>
+        <h1>Cookies</h1>
+        <p>Cookies are a common feature used on almost all websites. A cookie is a small text file that is downloaded
+            onto a device to allow a website to recognise it and to store some information about your preferences.</p>
+        <p>We use cookies for different reasons, including for managing your shopping cart, measuring your visits on our
+            sites, recognising and remembering your preferences and showing you personalised ads. Learn more about
+            cookies and how we use them.</p>
+        <div class="privacy-banner__form-container "></div>
+        <div class="privacy-banner__form-announcement" style="position: absolute;width:1px;height:1px;opacity:0;"
+            role="alert"></div>
+        </main>
+</body>
+
 </html>
