+++ packages/cookie-banner/package.json
@@ -23,9 +23,9 @@
   ],
   "scripts": {
     "build": "microbundle --name CookieBanner",
-    "test": "jest --coverage",
-    "dev": "webpack-dev-server --config example/webpack.config.js",
-    "prod": "webpack --config example/webpack.config.js --mode production",
+    "test": "jest --coverage & npx playwright test",
+    "dev": "webpack-dev-server --config tools/webpack.config.js",
+    "prod": "webpack --config tools/webpack.config.js --mode production",
     "prepublish": "npm run -s build"
   },
   "gitHead": "042ec0cbf12848b50c5962a5c8ab139cfb04df39"
