+++ packages/cookie-banner/tools/webpack.config.js
